project1_rest/repository/requirements.txt
Django==2.0.4
django-filter==1.1.0
django-extensions==2.0.7
djangorestframework==3.8.2
pytz==2018.4
six==1.11.0
